Updated Readme
Updated to give credit to the respected authors of the code. Several modifications were made in regards to styling, variable names, and some functions may be added to the code as well. There may also be tests that are written to compliment the code that is displayed for research purposes.